+++ IG_Sorter_Qt/src/ui/FlowLayout.cpp
@@ -1,0 +1,124 @@
+#include "ui/FlowLayout.h"
+#include <QWidget>
+
+FlowLayout::FlowLayout(QWidget* parent, int margin, int hSpacing, int vSpacing)
+    : QLayout(parent), m_hSpace(hSpacing), m_vSpace(vSpacing)
+{
+    setContentsMargins(margin, margin, margin, margin);
+}
+
+FlowLayout::FlowLayout(int margin, int hSpacing, int vSpacing)
+    : m_hSpace(hSpacing), m_vSpace(vSpacing)
+{
+    setContentsMargins(margin, margin, margin, margin);
+}
+
+FlowLayout::~FlowLayout() {
+    QLayoutItem* item;
+    while ((item = takeAt(0)))
+        delete item;
+}
+
+void FlowLayout::addItem(QLayoutItem* item) {
+    itemList.append(item);
+}
+
+int FlowLayout::horizontalSpacing() const {
+    if (m_hSpace >= 0) {
+        return m_hSpace;
+    }
+    return smartSpacing(QStyle::PM_LayoutHorizontalSpacing);
+}
+
+int FlowLayout::verticalSpacing() const {
+    if (m_vSpace >= 0) {
+        return m_vSpace;
+    }
+    return smartSpacing(QStyle::PM_LayoutVerticalSpacing);
+}
+
+int FlowLayout::count() const {
+    return itemList.size();
+}
+
+QLayoutItem* FlowLayout::itemAt(int index) const {
+    if (index >= 0 && index < itemList.size())
+        return itemList.at(index);
+    return nullptr;
+}
+
+QLayoutItem* FlowLayout::takeAt(int index) {
+    if (index >= 0 && index < itemList.size())
+        return itemList.takeAt(index);
+    return nullptr;
+}
+
+Qt::Orientations FlowLayout::expandingDirections() const {
+    return Qt::Horizontal;
+}
+
+bool FlowLayout::hasHeightForWidth() const {
+    return true;
+}
+
+int FlowLayout::heightForWidth(int width) const {
+    return doLayout(QRect(0, 0, width, 0), true);
+}
+
+void FlowLayout::setGeometry(const QRect& rect) {
+    QLayout::setGeometry(rect);
+    doLayout(rect, false);
+}
+
+QSize FlowLayout::sizeHint() const {
+    return minimumSize();
+}
+
+QSize FlowLayout::minimumSize() const {
+    QSize size;
+    for (const auto* item : itemList)
+        size = size.expandedTo(item->minimumSize());
+
+    const QMargins margins = contentsMargins();
+    size += QSize(margins.left() + margins.right(), margins.top() + margins.bottom());
+    return size;
+}
+
+int FlowLayout::doLayout(const QRect& rect, bool testOnly) const {
+    int left, top, right, bottom;
+    getContentsMargins(&left, &top, &right, &bottom);
+    QRect effectiveRect = rect.adjusted(+left, +top, -right, -bottom);
+    int x = effectiveRect.x();
+    int y = effectiveRect.y();
+    int lineHeight = 0;
+
+    for (auto* item : itemList) {
+        QSize spaceSize = item->sizeHint();
+        int nextX = x + spaceSize.width() + horizontalSpacing();
+        if (nextX - horizontalSpacing() > effectiveRect.right() && lineHeight > 0) {
+            x = effectiveRect.x();
+            y = y + lineHeight + verticalSpacing();
+            nextX = x + spaceSize.width() + horizontalSpacing();
+            lineHeight = 0;
+        }
+
+        if (!testOnly)
+            item->setGeometry(QRect(QPoint(x, y), spaceSize));
+
+        x = nextX;
+        lineHeight = qMax(lineHeight, spaceSize.height());
+    }
+    return y + lineHeight - rect.y() + bottom;
+}
+
+int FlowLayout::smartSpacing(QStyle::PixelMetric pm) const {
+    QObject* parent = this->parent();
+    if (!parent) {
+        return -1;
+    } else if (parent->isWidgetType()) {
+        QWidget* pw = static_cast<QWidget*>(parent);
+        return pw->style()->pixelMetric(pm, nullptr, pw);
+    } else {
+        return static_cast<QLayout*>(parent)->spacing();
+    }
+}