+++ IG_Sorter_Qt/src/ui/MenuScreen.h
@@ -22,4 +22,7 @@ private:
     QLabel*     m_configStatusLabel;
     QPushButton* m_startButton;
     QPushButton* m_settingsButton;
+
+private slots:
+    void openSourceFolder();
 };